WIP, added a MAVERange64 which is NSRange but always with uint64_t values, and start converting values

+++ MaveSDK/Utils/MAVEMerkleTree/MAVEMerkleTreeDataEnumerator.m
@@ -29,7 +29,7 @@
     return self._nextObject;
 }
 
-- (NSUInteger)keyForNextObject {
+- (uint64_t)keyForNextObject {
     return [[self peekAtNextObject] merkleTreeDataKey];
 }
